Beoordelen met decimalen, en naar boven afronden op een plek
Een cijfer is nu een schuif van 1 tot 10 in stappen van een tiende, met de waarde groot ernaast en een plus- en minknop voor het laatste tiende. Een trainer denkt in een zeven, maar wel een goeie; met hele cijfers moest hij kiezen tussen 7 en 8 en verdween precies het verschil dat hij zag.

De cijfertoetsen 1-9 en 0 blijven werken. Op een laptop is tikken sneller dan slepen, en dat rapport moet in dertig seconden klaar; de schuif is er voor precisie, niet in plaats van snelheid.

Op een telefoon: 44 pixels hoog, een grijper van 32, en touch-action none. Zonder dat laatste scrollt de pagina mee zodra je verticaal afwijkt en springt het cijfer terug.

report_scores.score heeft nu een decimaal en category_ratings bewaart de doorgerekende waarde ongeafrond -- daar hangt de voortgang aan, en bij afronden per stap telt een halve punt groei als een hele. Naar buiten wordt altijd naar boven afgerond: 74,5 wordt 75, in het voordeel van het kind.

Die afronding staat op een plek: CalculatePlayerCard::afronden(). Zodra er twee manieren van afronden in de app zitten laat het ene scherm 74 zien waar het andere 75 zegt, en dan gelooft niemand het meer. De voortgang, de maandsamenvatting en het dashboard gebruiken hem nu ook.

Verder op verzoek: het aandacht-blok is weg te klikken. Wat er wordt opgeslagen is een vingerafdruk van de inhoud, niet verborgen ja of nee -- verandert er iets, dan komt het blok terug. Voorgoed weg zou betekenen dat een school later niet weet dat er zeven rekeningen openstaan omdat iemand ooit op een kruisje drukte.

+++ app/Models/ReportScore.php
@@ -24,7 +24,9 @@ class ReportScore extends Model
     {
         return [
             'category' => ReportCategory::class,
-            'score' => 'integer',
+            // Eén decimaal: een trainer denkt in "een zeven, maar wel een
+            // goeie", en dat is een 7,4.
+            'score' => 'float',
         ];
     }
